requirements.txt
numpy==1.26.4
pandas==2.2.0
pycryptodome==3.16.0
requests==2.31.0
setuptools==65.6.3
streamlit==1.31.1
plotly==5.18.0
python-dotenv==1.0.1
websocket-client==1.7.0
schedule
